requirements.txt
python-telegram-bot==21.7
web3==7.6.0
python-dotenv==1.0.1
APScheduler==3.11.0
requests==2.32.3
